replace low level IPC with interprocess
