+++ crates/rpc/ipc/Cargo.toml
@@ -31,9 +31,6 @@ thiserror.workspace = true
 futures-util = "0.3.30"
 interprocess = { version = "1.2.1", features = ["tokio_support"] }
 
-[target.'cfg(windows)'.dependencies]
-windows-sys = { version = "0.52.0", features = ["Win32_Foundation"] }
-
 [dev-dependencies]
 tokio-stream = { workspace = true, features = ["sync"] }
 reth-tracing.workspace = true
